feat(sha256impl): prepare for new SHA256 implementations
The pycrypto SHA256 probably isn't the best/fastest implementation,
and the way it was included, as well as the shape of the SHA256
class etc., was somewhat messy.

Rip out the implementation to pycrypto_SHA256.h/.cpp, in preparation for
supporting other implementations.

Still a fair bit of work to do :)