feat(sha256impl): dynamic creation of concreate sha256 implementation
The sha256 class is now a virtual base class, with SHA256PyCrypto and
SHA256Sodium concrete implementations - Sodium an empty skeleton class
for now.

A Core class is introduced, which holds a vector of NamedHashCreator,
which will be used to dynamically create the selected hasher. Most of
the code now dynamically creates the hasher instance.

Need to benchmark once everything is dynamic, so we can see how much
performance is impaired by having to do virtual calls. If it's bad,
perhaps restructure the class hierarchy – perhaps template the Hash
class on the concrete implementation?